// types.ts
export interface ExpertProfile {
  name: string;
  role: string;
  location: string;
  whatsappLink: string;
  instagramLink: string;
  email: string;
  lipedemaLink: string;
}

export interface Feature {
  id: string;
  title: string;
  description: string;
  icon: string;
}

export interface ImageAsset {
  id: string;
  url: string;
  alt: string;
}